+++ nautilus_core/adapters/src/databento/python/live.rs
@@ -27,8 +27,7 @@ use nautilus_core::{
     time::{get_atomic_clock_realtime, AtomicTime, UnixNanos},
 };
 use nautilus_model::{
-    data::{delta::OrderBookDelta, deltas::OrderBookDeltas, Data},
-    ffi::data::deltas::OrderBookDeltas_API,
+    data::{delta::OrderBookDelta, deltas::OrderBookDeltas, deltas::OrderBookDeltas_API, Data},
     identifiers::{instrument_id::InstrumentId, symbol::Symbol, venue::Venue},
     python::data::data_to_pycapsule,
 };
