+++ nautilus_core/model/src/python/data/deltas.rs
@@ -23,8 +23,11 @@ use nautilus_core::time::UnixNanos;
 use pyo3::{prelude::*, pyclass::CompareOp, types::PyCapsule};
 
 use crate::{
-    data::{delta::OrderBookDelta, deltas::OrderBookDeltas, Data},
-    ffi::data::deltas::OrderBookDeltas_API,
+    data::{
+        delta::OrderBookDelta,
+        deltas::{OrderBookDeltas, OrderBookDeltas_API},
+        Data,
+    },
     identifiers::instrument_id::InstrumentId,
     python::common::PY_MODULE_MODEL,
 };
